Refactor form UI controller
